Extract result into its own file (#1255)
